Python_Bootcamp.Team_00/src/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.0
frozenlist==1.4.1
idna==3.7
interchange==2021.0.4
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
monotonic==1.6
multidict==6.0.5
networkx==3.3
numpy==2.0.0
nx-altair==0.1.6
packaging==24.1
pandas==2.2.2
pansi==2020.7.3
pillow==10.3.0
py2neo==2021.2.4
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
rpds-py==0.18.1
six==1.16.0
soupsieve==2.5
toolz==0.12.1
types-html5lib==1.1.11.20240228
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
